Refatoração do empilhaOperandoTipo

+++ trunk/ClassLoader/src/FuncoesBasicas.c
@@ -182,7 +182,7 @@ int jumpback(AmbienteExecucao *ae, int n_return) {
 		for (i = 0; i < n_return; i++) {
 			pilhaoperandos = desempilhaOperando(frame);
 			empilhaOperando(ae->pFrame, pilhaoperandos->tipo[pilhaoperandos->sp],
-					&(pilhaoperandos->elementos[pilhaoperandos->sp]));
+					(int)&(pilhaoperandos->elementos[pilhaoperandos->sp]));
 		}
 
 		return 0;
